+++ src/utils/geo_plot_helper.py
@@ -1,7 +1,9 @@
+#%%
 import sys
+import numpy as np
 import matplotlib.pyplot as plt
 import geopandas as gpd
-from haversine import haversine
+from haversine import haversine, Unit
 import math
 
 sys.path.append('/home/pcl/traffic/map_factory')
@@ -10,11 +12,73 @@ plt.rcParams['font.sans-serif'] = ['Microsoft YaHei']
 from ImageRelatedProcess import clip_background, merge_tiles
 import GoogleMapTile_V3 as tile
 
+ 
+def calculate_zoom(w, s, e, n, min_=12, max_=18):
+    # Note: we adopt the caculate zoom logic from contextily (https://contextily.readthedocs.io/en/latest/index.html)
+    """Automatically choose a zoom level given a desired number of tiles.
 
-def adaptive_zoom_level(distance, max_, min_):
-    # TODO
+    .. note:: all values are interpreted as latitude / longitutde.
+
+    Parameters
+    ----------
+    w : float
+        The western bbox edge.
+    s : float
+        The southern bbox edge.
+    e : float
+        The eastern bbox edge.
+    n : float
+        The northern bbox edge.
+
+    Returns
+    -------
+    zoom : int
+        The zoom level to use in order to download this number of tiles.
+    """
+    # Calculate bounds of the bbox
+    lon_range = np.sort([e, w])[::-1]
+    lat_range = np.sort([s, n])[::-1]
+
+    lon_length = np.subtract(*lon_range)
+    lat_length = np.subtract(*lat_range)
+
+    # Calculate the zoom
+    zoom_lon = np.ceil(np.log2(360 * 2.0 / lon_length))
+    zoom_lat = np.ceil(np.log2(360 * 2.0 / lat_length))
+    zoom = int(np.max([zoom_lon, zoom_lat]))
     
-    return    
+    if zoom > max_:
+        return max_
+    if zoom < min_:
+        return min_
+    
+    return zoom
+
+
+def adaptive_zoom_level(w, s, e, n, ax, factor=10, max_=19, min_=10 ):
+    dis = haversine((s, w), (n, e), unit=Unit.METERS)
+    scale_dict = { i: 40076000/256 / np.power(2, i) for i in range(0, 23)}
+
+    x, y = ax.get_figure().get_size_inches()
+    diagonal_pixels = np.sqrt(x*x + y*y ) * ax.get_figure().get_dpi()
+    act_scale = dis * factor / diagonal_pixels
+
+    z = 0
+    for key, val in scale_dict.items():
+        if act_scale < val:
+            continue
+        z = key
+        break
+    z = z- 1
+
+    if z <= min_:
+        return min_
+    if z >= max_:
+        return max_
+    
+    return z
+
+
 
 def map_visualize(df: gpd.GeoDataFrame, 
                   lyrs='s', 
@@ -65,11 +129,8 @@ def map_visualize(df: gpd.GeoDataFrame,
         [x0, y0, x1, y1] = [x0-(x1-x0) * scale, y0+(y0-y1) * scale,
                             x1+(x1-x0) * scale, y1-(y0-y1) * scale]
 
-    zoom = 15 - int(math.log2(haversine((x0, y1), (x1, y0))/3))
-    # print([x0, x1], [y0, y1], haversine((x0, y1), (x1, y0))/3)
-
-    # warming: if zoom big than 19 then there will be somthing wrong
-    zoom = 19 if zoom > 19 else zoom
+    zoom = calculate_zoom(x0, y0, x1, y1)
+    # zoom = adaptive_zoom_level(x0, y0, x1, y1, ax)
 
     img = tile.Tiles()
     f_lst, img_bbox = img.get_tiles_by_bbox([x0, y1, x1, y0], zoom, lyrs)
@@ -92,10 +153,13 @@ def map_visualize(df: gpd.GeoDataFrame,
 
 
 
-
 if __name__ == '__main__':
     from shapely.geometry import LineString
     line = LineString([(113.932686, 22.583023), (113.932679, 22.583161), (113.932679, 22.583221), (113.932654, 22.583295)])
     df = gpd.GeoDataFrame( [{"geometry": line}] )
-    map_visualize(df)
+    _, ax = map_visualize(df)
     
+    # daptive_zoom_level(*line.bounds, ax)
+    adaptive_zoom_level(*line.bounds, ax)
+    # %%
+    calculate_zoom(*line.bounds)